login works

+++ src/app/auth/guards/auth-guard.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
-import {Router} from '@angular/router';
-import {TokenService} from '../services/token.service';
+import { Router } from '@angular/router';
+import { TokenService } from '../services/token.service';
 
 @Injectable({
   providedIn: 'root'
@@ -13,14 +13,14 @@ export class AuthGuardService {
     // @ts-ignore
     // tslint:disable-next-line:max-line-length
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean {
-        /*if (!this.tokenService.getToken()) {// User not authenticated properly
+        if (!this.tokenService.getToken()) {// User not authenticated properly
             // Not authenticated properly
             this.tokenService.removeToken();
             this.router.navigate(['/login']);
             return false;
         } else {
             return true;
-        }*/
+        }
         return true;
     }
 }
